Data from the Box Lunch Study

This dataset was originally made public in the R package visTree (Venkatasubramaniam and Wolfson, 2018). The covariates of interest are laboratory- and questionairre-based measures of eating habits and attitudes towards food. The response is the average number of calories consumed in a 24 hour period. This data comes from the baseline measurements of the Box Lunch Study (French et al., 2014)

blsdata

Format

trt

Treatment

sex

Sex

bmi0

BMI

skcal

Snacking kilo calories.

srvgfv0

Serving size of fruits and vegetables.

srvgssb

Serving size of beverages

kcal24h0

Response variable; kcals consumed in 24 hours.

disinhibition

wanting

hunger

liking

resteating

edeq01

edeq02

edeq13

edeq14

edeq15

edeq22

edeq23

edeq25

edeq26

cdrsbody0

Body image

weighfreq0

Weighing frequency

freqff

Fast food frequency

age

Age

rrvfood

Relative reinforcement of food

A data frame with 226 rows and 26 variables

References

Ashwini Venkatasubramaniam and Julian Wolfson (2018). visTree: Visualization of Subgroups for Decision Trees. R package version 0.8.1. https://CRAN.R-project.org/package=visTree

French SA, Mitchell NR, Wolfson J, Harnack LJ, Jeffery RW, Gerlach AF, Blundell JE, Pentel PR. Portion size effects on weight gain in a free living setting. Obesity. 2014;22(6):1400–5.

Examples

data(blsdata)