compute selective confidence intervals

This function computes the selective confidence intervals. Code modified from the Outference package, available at https://github.com/shuxiaoc/outference.

computeCI(v, y, sigma = NULL, truncation, alpha = 0.05)

Arguments

v
the contrast vector that defines the parameter of interest

y
the observed response vector

sigma
The known noise standard deviation. If unknown, we recommend a conservative estimate. If it is left blank, we automatically use a conservative estimate.

truncation,
the truncation set for the statistic v'y. Computes a confidence interval for the mean of a truncated normal distribution.

alpha,
the significance level.

Value

This function returns a vector of lower and upper confidence limits.