Pass in an rpart tree and also a node number. The nodes are numbered strangely in rpart. If no node numbers are passed in, a list of all branches in the tree will be returned. Pass in a tree and a node number (node number like the type that comes up in rpart.plot!!!)

Pass in an rpart tree and also a node number. The nodes are numbered strangely in rpart. If no node numbers are passed in, a list of all branches in the tree will be returned. Pass in a tree and a node number (node number like the type that comes up in rpart.plot!!!)

getBranch(tree, nn = NULL)

Arguments

tree
An rpart object.

nn
A node number that corresponds to a valid node in ``tree``. The list of valid node numbers can be obtained with ``row.names(tree$frame)`` or by plotting ``tree`` with ``treeval.plot`` and ``nn=TRUE``. The node number can be passed in as either a character string or an integer.

Value

Either a single branch (which is a vector of splits) or (if nn=NULL), a list of all branches in the tree.